Added item.js file
