+++ app.js
@@ -4,9 +4,10 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
 var authMiddleware = require('./middleware/auth');
+var indexRouter = require('./routes/index');
+var itemsRouter = require('./routes/items');
+var usersRouter = require('./routes/users');
 
 var app = express();
 
@@ -20,7 +21,8 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/api/items', authMiddleware, indexRouter);
+app.use('/', indexRouter);
+app.use('/api/items', authMiddleware, itemsRouter);
 app.use('/users', usersRouter);
 
 // catch 404 and forward to error handler
